requirements.txt
# database
db-sqlite3

# discord API
discord.py[voice]
python-dotenv

# youtube API 
yt-dlp
google-api-python-client
google-auth-oauthlib 
google-auth-httplib2

# github
gitpython

# data analysis / web scraping
selenium
nltk
regex
pandas

# chess cog
python-chess
cairosvg

# x/twitter API
tweepy
standard-imghdr

# codeforces API
CodeforcesApiPy

# pc control
keyboard
mouse
pyautogui
pywin32

# deployment
pyinstaller

# other
urllib3
july
audioop-lts
pillow